feat(multiselect): get input props
