+++ packages/chakra-components/src/MultiSelect/useComboBox.js
@@ -1,16 +1,11 @@
-import { useState, useEffect } from "react";
+import { useState, useEffect, useRef } from "react";
 
-const defaultFilter = (options, input) => {
-  if (input) {
-    return options.filter(option =>
-      option.label.toLowerCase().includes(input.toLowerCase()),
-    );
-  } else {
-    return options;
-  }
-};
+export const useComboBox = ({ options, initialValues, isMulti }) => {
+  const listRef = useRef(null);
+  const inputRef = useRef(null);
+  const multiSelectRef = useRef(null);
+  const popperRef = useRef(null);
 
-export const useComboBox = ({ options, initialValues, isMulti, listRef }) => {
   let _initialValues = [];
 
   if (initialValues) {
@@ -30,6 +25,16 @@ export const useComboBox = ({ options, initialValues, isMulti, listRef }) => {
   const [selectedOptions, setSelectedOptions] = useState([]);
   const [notSelectedOptions, setNotSelectedOptions] = useState(options);
   const [filteredOptions, setFilteredOptions] = useState(notSelectedOptions);
+
+  const defaultFilter = (options, input) => {
+    if (input) {
+      return options.filter(option =>
+        option.label.toLowerCase().includes(input.toLowerCase()),
+      );
+    } else {
+      return options;
+    }
+  };
 
   useEffect(() => {
     setSelectedOptions(
@@ -67,6 +72,237 @@ export const useComboBox = ({ options, initialValues, isMulti, listRef }) => {
     }
   }, [isMulti, filteredOptions, values]);
 
+  const getWrapperProps = () => {
+    return {
+      tabIndex: -1,
+      onClick: e => {
+        if (!isFocused) {
+          inputRef.current.focus();
+        }
+
+        if (inputIsHidden) {
+          setInputIsHidden(false);
+        }
+
+        if (isOpen) {
+          setIsOpen(false);
+          setInputValue("");
+        } else {
+          setIsOpen(true);
+
+          if (!isMulti) {
+            if (!values.length) {
+              setFocusedOptionIndex(0);
+            } else {
+              const selectedOption = filteredOptions.find(
+                option => option.value === values[0],
+              );
+
+              const selectedIndex = filteredOptions.indexOf(selectedOption);
+
+              if (selectedIndex !== -1) {
+                setFocusedOptionIndex(selectedIndex);
+              }
+            }
+          } else {
+            setFocusedOptionIndex(0);
+          }
+        }
+      },
+    };
+  };
+
+  const getInputProps = () => {
+    return {
+      onChange: event => {
+        setInputValue(event.currentTarget.value);
+
+        if (inputIsHidden) {
+          setInputIsHidden(false);
+        }
+
+        if (focusedOptionIndex !== 0) {
+          setFocusedOptionIndex(0);
+        }
+
+        if (!isOpen) {
+          setIsOpen(true);
+        }
+      },
+      onFocus: event => {
+        if (!isFocused) {
+          setIsFocused(true);
+        }
+
+        if (inputIsHidden) {
+          setInputIsHidden(false);
+        }
+      },
+      onKeyDown: event => {
+        const count = filteredOptions.length;
+        let nextIndex;
+
+        if (event.key === "ArrowDown") {
+          if (filteredOptions.length) {
+            if (inputIsHidden) {
+              setInputIsHidden(false);
+            }
+
+            if (!isOpen) {
+              setIsOpen(true);
+
+              if (!isMulti) {
+                if (!values.length) {
+                  setFocusedOptionIndex(0);
+                } else {
+                  const selectedOption = filteredOptions.find(
+                    option => option.value === values[0],
+                  );
+
+                  const selectedIndex = filteredOptions.indexOf(selectedOption);
+
+                  if (selectedIndex !== -1) {
+                    setFocusedOptionIndex(selectedIndex);
+                  }
+                }
+              } else {
+                setFocusedOptionIndex(0);
+              }
+
+              return;
+            }
+
+            nextIndex = (focusedOptionIndex + 1) % count;
+            setFocusedOptionIndex(nextIndex);
+          }
+        }
+
+        if (event.key === "ArrowUp") {
+          if (filteredOptions.length) {
+            if (inputIsHidden) {
+              setInputIsHidden(false);
+            }
+
+            if (!isOpen) {
+              setIsOpen(true);
+
+              if (!isMulti) {
+                if (!values.length) {
+                  setFocusedOptionIndex(filteredOptions.length - 1);
+                } else {
+                  const selectedOption = filteredOptions.find(
+                    option => option.value === values[0],
+                  );
+
+                  const selectedIndex = filteredOptions.indexOf(selectedOption);
+
+                  if (selectedIndex !== -1) {
+                    setFocusedOptionIndex(selectedIndex);
+                  }
+                }
+              } else {
+                setFocusedOptionIndex(filteredOptions.length - 1);
+              }
+
+              return;
+            }
+
+            nextIndex = (focusedOptionIndex - 1 + count) % count;
+            setFocusedOptionIndex(nextIndex);
+          }
+        }
+
+        if (event.key === "Enter") {
+          if (filteredOptions.length) {
+            const focusedOption = filteredOptions[focusedOptionIndex];
+
+            if (focusedOption.disabled) {
+              return;
+            }
+
+            if (isOpen) {
+              if (!isMulti) {
+                setValues([focusedOption.value]);
+
+                if (!inputIsHidden) {
+                  setInputIsHidden(true);
+                }
+              } else {
+                setValues(oldOptions => {
+                  if (oldOptions.includes(focusedOption.value)) {
+                    return oldOptions;
+                  }
+
+                  return [...oldOptions, focusedOption.value];
+                });
+              }
+            }
+
+            setInputValue("");
+
+            if (isOpen) {
+              setIsOpen(false);
+            }
+          }
+        }
+
+        if (event.key === "Escape") {
+          if (isOpen) {
+            setIsOpen(false);
+          }
+        }
+
+        if (event.key === "Backspace") {
+          if (!inputValue) {
+            if (inputIsHidden) {
+              setInputIsHidden(false);
+            }
+
+            if (!isMulti) {
+              setValues([]);
+            } else {
+              if (values.length) {
+                const newValues = values.slice(0, values.length - 1);
+                setValues(newValues);
+              }
+            }
+          }
+        }
+      },
+      onBlur: event => {
+        if (!isOpen && !multiSelectRef.current.contains(event.relatedTarget)) {
+          if (isFocused) {
+            setIsFocused(false);
+          }
+
+          setInputValue("");
+
+          return;
+        }
+
+        if (
+          isOpen &&
+          !popperRef.current.contains(event.relatedTarget) &&
+          !multiSelectRef.current.contains(event.relatedTarget)
+        ) {
+          if (isFocused) {
+            setIsFocused(false);
+          }
+
+          if (isOpen) {
+            setIsOpen(false);
+          }
+
+          setInputValue("");
+
+          return;
+        }
+
+        inputRef.current.focus();
+      },
+    };
+  };
+
   return {
     values,
     setValues,
@@ -83,5 +319,11 @@ export const useComboBox = ({ options, initialValues, isMulti, listRef }) => {
     selectedOptions,
     filteredOptions,
     setFilteredOptions,
+    multiSelectRef,
+    inputRef,
+    popperRef,
+    listRef,
+    getWrapperProps,
+    getInputProps,
   };
 };
